+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Additions to the :doc:`core-object-methods`.
 
 * Added the :py:meth:`khoros.core.Khoros.Board.get_message_count` method.
 * Added the :py:meth:`khoros.core.Khoros.Board.get_all_messages` method.
+* Added the ability to define the logging level when instantiating the core object.
 
 Primary Modules
 ---------------